SSL: Improve X.509 certificate loading from openssl engine with file source
In some cases key pair and its certificate are stored in engine but
intermediate certificates are not.

Add support for loading from the file system the rest of the chain.

As a side effect of using space as separator and when loading additional
certificates for chain parts from file system it cannot handle spaces in
the path.

Example configuration with libp11's pkcs11 engine:

  ssl_certificate      "engine:pkcs11:pkcs11:token=mytoken;object=mykey
                        /etc/ssl/certs/intermediate-ca.pem";
  ssl_certificate_key  "engine:pkcs11:pkcs11:token=mytoken;object=mykey?pin-value=mypin";

Signed-off-by: Vesa Jääskeläinen <vesa.jaaskelainen@vaisala.com>